% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_winner.R
\name{county_winner}
\alias{county_winner}
\title{County Winner}
\usage{
county_winner(Year = 2016, states = c())
}
\arguments{
\item{Year}{United States Presidential Election year starting in 2000.}

\item{states}{region to plot.}
}
\value{
a ggplot map object.
}
\description{
county_winner displays a map of presidential election results by party on the county level
}
\examples{
county_winner(Year = 2008)
county_winner(Year = 2000, states=c("texas", "oklahoma"))
}
