% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_growth.R
\name{covid_growth}
\alias{covid_growth}
\title{Covid Growth Over Time}
\usage{
covid_growth(
  states = c(),
  value,
  sum = FALSE,
  df_print = FALSE,
  interact = TRUE,
  ...
)
}
\arguments{
\item{states}{region to analyze.}

\item{value}{covid statistic to analyze.}

\item{sum}{take sum of all cases over states or treat states individually.}

\item{df_print}{print a data frame or not.}

\item{interact}{whether additional information will display when user hovers over image (use ggplot or plotly)}

\item{...}{pass other arguments}
}
\value{
a plotly or ggplot object.
}
\description{
covid_growth streamlines the process for plotting a line graph to analyze
longitudinal changes in a user-selected variable over time. Variable options include
total cases ("tot_cases"), new cases ("new_cases"), total deaths ("tot_death"), and
new deaths ("new_death").
}
\examples{
covid_growth(states="colorado", df_print=FALSE, sum=FALSE)
covid_growth(states=c("colorado", "nevada"))
covid_growth(states=c("colorado", "nevada", "florida", "california"), df_print=FALSE, sum=FALSE)
covid_growth(states=c("colorado", "nevada", "florida", "california"), interact=FALSE)
}
\seealso{
ggplot
}
