% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_map.R
\name{covid_map}
\alias{covid_map}
\title{Covid Map}
\usage{
covid_map(value, Date = "2020-12-02", pal = "YlOrRd", ...)
}
\arguments{
\item{value}{covid statistic to analyze.}

\item{Date}{a date in form "YYYY-MM-DD".  --> CHOOSE DEFAULT}

\item{pal}{a color palette.}

\item{...}{pass other arguments}
}
\value{
a ggplot map object.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
covid_map(Value="new_death", Date="2020-12-01")
}
