% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5 covid_update_func.R
\name{covid_update}
\alias{covid_update}
\title{covid_update}
\usage{
covid_update()
}
\value{
a data frame
}
\description{
covid_update downloads and cleans the latest COVID-19 data from the CDC website.
}
\examples{
library(USevolve)
covid_data <- covid_update()
}
