% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1 elect_count_func.R
\name{elect_count}
\alias{elect_count}
\title{Election Count}
\usage{
elect_count(level = "state", Year, states = c(), log = TRUE)
}
\arguments{
\item{level}{what region size to visualize (state or county).}

\item{Year}{year. For built in election data, year must align with US election years.}

\item{states}{list of states user wants to visualize.}

\item{log}{fill based on log of count.}
}
\value{
a ggplot map
}
\description{
elect_count creates a map of the United States with fill color based on the specified parameter.
}
\examples{
elect_count()
elect_count(data="county", states = c("connecticut", "rhode island"), log=F)

}
