% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4 time_growth_func.R
\name{time_growth}
\alias{time_growth}
\title{Growth Over Time}
\usage{
time_growth(data = covid, states = c(), sum = FALSE, df_print = TRUE)
}
\arguments{
\item{data}{a data frame. Default is Covid data.}

\item{states}{region to analyze.}

\item{sum}{take sum of all cases over states or treat states individually.}

\item{df_print}{print a data frame or not.}
}
\value{
a data frame and ggplot graph
}
\description{
time_growth streamlines the process for plotting longitudinal changes in
a variable over time.
}
\examples{
time_growth(states="colorado", df_print=F)
time_growth(states=c("colorado", "nevada"), sum=T)
time_growth(states=c("colorado", "nevada", "florida", "california"), df_print=F)
}
