% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_growth.R
\name{covid_growth}
\alias{covid_growth}
\title{Covid Growth Over Time}
\usage{
covid_growth(states = c(), value, sum = FALSE, df_print = FALSE, ...)
}
\arguments{
\item{states}{region to analyze.}

\item{value}{covid statistic to analyze.}

\item{sum}{take sum of all cases over states or treat states individually.}

\item{df_print}{print a data frame or not.}

\item{...}{pass other arguments}
}
\value{
a list (plotly object and htmlwidget) CHECK
}
\description{
covid_growth streamlines the process for plotting longitudinal changes in
a variable over time.
}
\examples{
covid_growth(states="colorado", df_print=FALSE, sum=FALSE)
covid_growth(states=c("colorado", "nevada"))
covid_growth(states=c("colorado", "nevada", "florida", "california"), df_print=FALSE, sum=FALSE)
}
