% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6 covid_map_func.R
\name{covid_map}
\alias{covid_map}
\title{Covid Map}
\usage{
covid_map(Value, Date = "2020-12-02", pal = "YlOrRd", ...)
}
\arguments{
\item{Value}{DESCRIPTION.}

\item{Date}{a date in form "YYYY-MM-DD".  --> CHOOSE DEFAULT}
}
\value{
a ggplot map object.
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
covid_map(Value="new_death", Date="2020-12-01")
}
