% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_map.R
\name{covid_map}
\alias{covid_map}
\title{Covid Map}
\usage{
covid_map(value, Date = "2020-12-02", pal = "YlOrRd")
}
\arguments{
\item{value}{covid statistic to analyze.}

\item{Date}{a date in form "YYYY-MM-DD".}

\item{pal}{a color palette.}
}
\value{
a leaflet object.
}
\description{
covid_map automatically loads the latest COVID-19 data from the CDC and plots a chloropleth
map of the United States based on the fill value the user specifies. Fill options include
total cases ("tot_cases"), new cases ("new_cases"), total deaths ("tot_death"), and
new deaths ("new_death").
}
\examples{
covid_map(value="new_death", Date="2020-12-01")
}
