% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4 covid_update_func.R
\name{covid_update}
\alias{covid_update}
\title{covid_update}
\usage{
covid_update()
}
\description{
covid_update downloads and cleans the latest COVID-19 data from the CDC website.
}
\keyword{internal}
