% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elect_count.R
\name{elect_count}
\alias{elect_count}
\title{Election Count}
\usage{
elect_count(
  level = "state",
  Year,
  states = c(),
  pal = "YlOrRd",
  log = TRUE,
  ...
)
}
\arguments{
\item{level}{what region size to visualize (state or county).}

\item{Year}{year. For built in election data, year must align with US election years.}

\item{states}{list of states user wants to visualize.}

\item{pal}{a color palette.}

\item{log}{fill based on log of count.}

\item{...}{pass other arguments}
}
\value{
a ggplot map object.
}
\description{
elect_count creates a map of the United States with fill color based on the specified parameter.
}
\examples{
elect_count()
elect_count(level="county", states = c("connecticut", "rhode island"), log=FALSE)

}
